% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_field_values}
\alias{annotate_field_values}
\title{Annotate Field Values}
\usage{
annotate_field_values(data, ...)
}
\arguments{
\item{data}{dataframe returned from a Metathesaurus query}

\item{...}{(optional) Columns in the data to annotate. If missing, all available annotations are provided for every column.}
}
\description{
These functions add extra fields to a dataframe that annotate the field value meaning with what is provided in the UMLS Web Documentation.
}
\details{
These annotations are derived from \code{\link{get_field_value_annotations}}
}
\seealso{
\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[purrr]{reduce}}
}
