% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{annotate_sab}
\alias{annotate_sab}
\title{Annotate the SAB Field}
\usage{
annotate_sab(data)
}
\description{
Annotate the SAB Field
}
\details{
The annotations used in this function are derived from \code{\link{get_sab_annotation}}.
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{mutate-joins}}
}
