% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotation_functions}
\alias{annotation_functions}
\title{Field Value Annotation Functions}
\arguments{
\item{data}{dataframe returned from a Metathesaurus query}
}
\description{
These functions add extra fields to a dataframe that annotate the field value meaning with what is provided in the UMLS Web Documentation.
}
\keyword{internal}
