% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_to_omop_functions.R
\name{crosswalk_to_omop_functions}
\alias{crosswalk_to_omop_functions}
\title{Crosswalk To OMOP Concepts}
\description{
These functions provide avenues for crosswalking to the OMOP Vocabulary.
}
\keyword{internal}
