% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc_mrconso_field_definitions}
\alias{doc_mrconso_field_definitions}
\title{Get Field Definitions for the MRCONSO Table}
\usage{
doc_mrconso_field_definitions()
}
\description{
Get Field Definitions for the MRCONSO Table
}
\seealso{
\code{\link[tibble]{tribble}}
}
