% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc_readable_table_names}
\alias{doc_readable_table_names}
\title{List the Human Readable Names for the Tables}
\usage{
doc_readable_table_names()
}
\description{
List the Human Readable Names for the Tables
}
\seealso{
\code{\link[tibble]{tribble}}
}
