% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_scraping_functions.R
\name{documentation_scraping_functions}
\alias{documentation_scraping_functions}
\title{Scrape UMLS Documentation into Tables}
\description{
These functions scrape the UMLS Web Documentation to write the \code{\link{mth_documentation_functions}}. They have been kept to enable reproducibility and other long-term maintenance efforts.
}
\keyword{internal}
