% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{field_annotation}
\alias{field_annotation}
\title{Annotate Field Values with Details in UMLS Documentation}
\usage{
field_annotation()
}
\description{
Get Maps Between Field-Value Pairs where the Value is abbreviated as show at \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html}{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/columns_data_elements.html}.
}
\seealso{
\code{\link[tibble]{tribble}}
}
