% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_nci.R, R/filter_for_preferred_term.R
\name{filter_for_nci}
\alias{filter_for_nci}
\title{Filter MRCONSO output for NCI}
\usage{
filter_for_nci(mrconso_df)

filter_for_nci(mrconso_df)
}
\description{
Filter MRCONSO output for NCI

Filter MRCONSO output for preferred_term
}
