% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{get_field_definitions}
\alias{get_field_definitions}
\title{Get All Available Field Definitions}
\usage{
get_field_definitions()
}
\description{
Sourced from \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/columns_data_elements.html}{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/columns_data_elements.html}
}
\seealso{
\code{\link[tibble]{tribble}}
}
