% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{get_fields_to_table_map}
\alias{get_fields_to_table_map}
\title{Map all the MTH Fields to the Table
 \code{\link[tibble]{tribble}}}
\usage{
get_fields_to_table_map()
}
\description{
Map all the MTH Fields to the Table
 \code{\link[tibble]{tribble}}
}
