% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{get_sab_annotation}
\alias{get_sab_annotation}
\title{Get Vocabulary Details to Annotate SAB Fields}
\usage{
get_sab_annotation()
}
\description{
Get Maps Between the SAB where the Vocabulary is referred to by Abbreviation. Expansion is sourced from \href{https://www.nlm.nih.gov/research/umls/sourcereleasedocs/index.html}{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/columns_data_elements.html}
}
\seealso{
\code{\link[tibble]{tribble}}
}
