% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_to_omop_functions.R
\name{get_semantic_type_to_domain_id}
\alias{get_semantic_type_to_domain_id}
\title{Crosswalk Between Semantic Type and OMOP Domain Id}
\usage{
get_semantic_type_to_domain_id()
}
\description{
Crosswalk Between Semantic Type and OMOP Domain Id
}
\details{
Note that this crosswalk is subjective and is constrained to the Observation, Device, Drug, Procedure, and Measurement domains only. If a Semantic Type is not applicable to OMOP, such as Fish and Reptile Semantic Types, they are given a Domain Id of NA. It is also important to note that the mappings between Semantic Type and OMOP Domain Id can be 1 to many.
}
\seealso{
\code{\link[tibble]{tribble}}
}
