% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihSite.R
\name{getAbbreviationInfo}
\alias{getAbbreviationInfo}
\title{Get Abbreviation Information}
\usage{
getAbbreviationInfo()
}
\value{
A list of data frames scraped from \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html}{Abbreviations}
}
\description{
Get all the tables that provide abbreviation information at the NIH site \href{https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html}{Abbreviations}
}
\seealso{
\code{\link[xml2]{read_xml}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_table}}
 \code{\link[purrr]{set_names}}
}
