% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nihSite.R
\name{getVocabularyInfo}
\alias{getVocabularyInfo}
\title{Get Source Vocabulary Information}
\usage{
getVocabularyInfo()
}
\value{
A data frame of the table found at  \href{https://www.nlm.nih.gov/research/umls/sourcereleasedocs/index.html}{UMLS Metathesaurus Vocabulary Documentation}.
}
\description{
This function returns basic information on all the source vocabularies found in the UMLS Metathesaurus from the NIH site for \href{https://www.nlm.nih.gov/research/umls/sourcereleasedocs/index.html}{UMLS Metathesaurus Vocabulary Documentation}.
}
\details{
This function assumes that the table of interest is found at https://www.nlm.nih.gov/research/umls/sourcereleasedocs/index.html. Future modifications may be needed to update the web address should it be changed.
}
