% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_fields}
\alias{list_fields}
\title{List Fields}
\usage{
list_fields(conn, schema = "mth", table)
}
\arguments{
\item{conn}{Postgres Connection}

\item{schema}{Schema of MTH instance}

\item{table}{MTH Table}
}
\description{
Return a valueset to see what the possible values are in a given Field
}
\details{
List the fields in a MTH Table.
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}
}
