% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_functions}
\alias{list_functions}
\title{List Database Elements}
\arguments{
\item{conn}{Postgres Connection}

\item{schema}{Schema of MTH instance}
}
\value{
Vector of 1 or more values
}
\description{
List details on the MTH database architecture
}
\keyword{internal}
