% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_code.R
\name{lookup_code}
\alias{lookup_code}
\title{Lookup a Code in MRCONSO Table}
\usage{
lookup_code(conn, code, sab, schema = "mth")
}
\arguments{
\item{conn}{Postgres Connection}

\item{schema}{Schema of MTH instance}
}
\value{
Resultset of 1 or more rows
}
\description{
Lookup the elements of a CUI
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}
}
\concept{code lookup}
