% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cui_functions.R
\name{lookup_cui_attributes}
\alias{lookup_cui_attributes}
\title{Lookup a CUI in the MRSAT Table}
\usage{
lookup_cui_attributes(conn, cui, schema = "mth")
}
\arguments{
\item{conn}{Postgres Connection}

\item{cui}{CUI}

\item{schema}{Schema of MTH instance}
}
\value{
Resultset of 1 or more rows
}
\description{
Lookup the elements of a CUI
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}

Other cui lookup: 
\code{\link{lookup_cui_inverse_rel}()},
\code{\link{lookup_cui_relationships}()},
\code{\link{lookup_cui_semantic_types}()},
\code{\link{lookup_cui_source_context}()},
\code{\link{lookup_cui_source_defs}()},
\code{\link{lookup_cui}()}
}
\concept{cui lookup}
