% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cui_functions.R
\name{lookup_cui_functions}
\alias{lookup_cui_functions}
\title{Lookup CUI Functions}
\arguments{
\item{conn}{Postgres Connection}

\item{cui}{CUI}

\item{schema}{Schema of MTH instance}
}
\value{
Resultset of 1 or more rows
}
\description{
Lookup the elements of a CUI
}
\keyword{internal}
