% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_inverse_rel}
\alias{lookup_inverse_rel}
\title{Lookup a CUI in MRREL Table}
\usage{
lookup_inverse_rel(conn, cui, schema = "mth")
}
\arguments{
\item{conn}{Postgres Connection}

\item{cui}{CUI}

\item{schema}{Schema of MTH instance}
}
\value{
Resultset of 1 or more rows
}
\description{
Lookup the elements of a CUI
}
\details{
This function sets the `cui` argument to CUI2in the MRREL Table. For the non-inverse relationship, see \code{\link{lookup_relationships}}
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}

Other cui lookup: 
\code{\link{lookup_attributes}()},
\code{\link{lookup_cui}()},
\code{\link{lookup_relationships}()},
\code{\link{lookup_semantic_types}()},
\code{\link{lookup_source_context}()},
\code{\link{lookup_source_defs}()}
}
\concept{cui lookup}
