% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_source_defs}
\alias{lookup_source_defs}
\title{Lookup a CUI in the MRDEF Table}
\usage{
lookup_source_defs(conn, cui, schema = "mth")
}
\arguments{
\item{conn}{Postgres Connection}

\item{cui}{CUI}

\item{schema}{Schema of MTH instance}
}
\value{
Resultset of 1 or more rows
}
\description{
Lookup the elements of a CUI
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}

Other cui lookup: 
\code{\link{lookup_attributes}()},
\code{\link{lookup_cui}()},
\code{\link{lookup_inverse_rel}()},
\code{\link{lookup_relationships}()},
\code{\link{lookup_semantic_types}()},
\code{\link{lookup_source_context}()}
}
\concept{cui lookup}
