% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconso_search_functions.R
\name{mrconso_search_functions}
\alias{mrconso_search_functions}
\title{Search MRCONSO Functions}
\arguments{
\item{conn}{Postgres connection}

\item{schema}{Metathesaurus schema}

\item{concept}{Concept to be searched}

\item{match_case}{If TRUE, both the MRCONSO STR column and the `concept` argument are converted to lowercase.}

\item{cache_only}{If TRUE, only the cache is loaded and database querying is skipped. If a cache did not exist for the query, a NULL will be returned.}

\item{skip_cache}{If TRUE, the cache is skipped altogether. The database is directly queried and the resultset is not cached.}

\item{override_cache}{If TRUE, the cache is not loaded, the database is directly queried and the resultset is cached, overwriting any existing cache that may have been present.}
}
\value{
Dataframe of 0 or more rows
}
\description{
These functions help search for mappings between an a concept and a Metathesaurus identifier in the MRCONSO Table. All queries are cached using the R.cache package for quicker retrieval of the more computationally expensive searches.
}
\seealso{
Other search mrconso: 
\code{\link{search_exact}()},
\code{\link{search_like}()},
\code{\link{search_string}()}
}
\concept{search mrconso}
