% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_file_description_functions.R
\name{MRXNS_ENG_def}
\alias{MRXNS_ENG_def}
\title{Normalized String Index (File = MRXNS_ENG.RRF)}
\usage{
MRXNS_ENG_def()
}
\description{
Normalized String Index (File = MRXNS_ENG.RRF)
}
\details{
There is one row in this table for each normalized string found in each unique English-language Metathesaurus string (ignoring upper-lower case). All English-language Metathesaurus entries have entries in the normalized string index. There are no normalized word indexes for other languages in this edition of the Metathesaurus.
}
\seealso{
\code{\link[tibble]{tribble}}
}
