% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_file_description_functions.R
\name{MRXNW_ENG_def}
\alias{MRXNW_ENG_def}
\title{Normalized Word Index (File = MRXNW_ENG.RRF)}
\usage{
MRXNW_ENG_def()
}
\description{
Normalized Word Index (File = MRXNW_ENG.RRF)
}
\details{
There is one row in this table for each normalized word found in each unique English-language Metathesaurus string. All English-language Metathesaurus entries have entries in the normalized word index. There are no normalized string indexes for other languages in the Metathesaurus.
}
\seealso{
\code{\link[tibble]{tribble}}
}
