% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{mth_documentation_functions}
\alias{mth_documentation_functions}
\title{Metathesaurus Documentation Functions}
\description{
Return the information avaible in the UML Web Documentation in dataframe format
}
