% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{mth_fields_to_tables}
\alias{mth_fields_to_tables}
\title{Map all the MTH Fields to the Table
 \code{\link[tibble]{tribble}}}
\usage{
mth_fields_to_tables()
}
\description{
Map all the MTH Fields to the Table
 \code{\link[tibble]{tribble}}
}
