% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_file_description_functions.R
\name{mth_file_description_functions}
\alias{mth_file_description_functions}
\title{Metathesaurus File Descriptions}
\description{
This family of functions returns the RRF file descriptions found at
\href{https://www.ncbi.nlm.nih.gov/books/NBK9685/}{https://www.ncbi.nlm.nih.gov/books/NBK9685/} as a tibble.
}
