% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cui.R
\name{query_cui}
\alias{query_cui}
\title{Queries cui in UMLS in MRCONSO table of metathesaurus}
\usage{
query_cui(cui, limit = NULL)
}
\arguments{
\item{limit}{LIMIT in the sql statement. If NULL, all records will be returned.}

\item{umls_cui}{C-prefixed CUI from metathesaurus}
}
\value{
resultset from SQL query
}
\description{
Queries cui in UMLS in MRCONSO table of metathesaurus
}
