% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_cui_attributes}
\alias{query_cui_attributes}
\title{Query for Attributes}
\usage{
query_cui_attributes(conn, cui, schema = "mth")
}
\description{
Query for Attributes
}
\seealso{
Other query: 
\code{\link{query_cui_semantic_types}()},
\code{\link{query_cui_source_context}()},
\code{\link{query_cui_source_definitions}()}
}
\concept{query}
