% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cui_level_functions.R
\name{query_cui_inverse_relationships}
\alias{query_cui_inverse_relationships}
\title{Query for source definition}
\usage{
query_cui_inverse_relationships(cui)
}
\description{
Query for source definition
}
