% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{query_cui_source_definitions}
\alias{query_cui_source_definitions}
\title{Query for source definitions}
\usage{
query_cui_source_definitions(conn, cui, schema = "mth")
}
\description{
Query for source definitions
}
\seealso{
Other query: 
\code{\link{query_cui_attributes}()},
\code{\link{query_cui_semantic_types}()},
\code{\link{query_cui_source_context}()}
}
\concept{query}
