% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cui_vector.R
\name{query_cui_vector}
\alias{query_cui_vector}
\title{Filters MRCONSO table for CUI values that exactly match values in the argument vector}
\usage{
query_cui_vector(vector, limit_per_cui = NULL)
}
\arguments{
\item{vector}{vector of CUIs for IN statement in SQL query}
}
\value{
resultSet where CUI IN vector argument
}
\description{
Filters MRCONSO table for CUI values that exactly match values in the argument vector
}
