% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_exact_phrase_match.R
\name{query_exact_phrase_match}
\alias{query_exact_phrase_match}
\title{Query based on search terms}
\usage{
query_exact_phrase_match(phrase, limit = NULL)
}
\description{
Query based on search terms
}
