% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nci_cl_cuis.R
\name{query_nci_cl_cuis}
\alias{query_nci_cl_cuis}
\title{Query NCI metathesaurus for all "CL" cuis}
\usage{
query_nci_cl_cuis(nci_code_cui_map_table = "nci_code_cui_map_201808")
}
\description{
Query NCI metathesaurus for all "CL" cuis
}
