% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nci_cui.R
\name{query_nci_cui}
\alias{query_nci_cui}
\title{Query for source definition}
\usage{
query_nci_cui(nci_cui, nci_code_cui_map_table = "nci_code_cui_map_201808")
}
\description{
Query for source definition
}
