% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nci_exact_phrase.R
\name{query_nci_exact_phrase}
\alias{query_nci_exact_phrase}
\title{Query NCI metathesaurus for all exact phrase matches}
\usage{
query_nci_exact_phrase(
  nci_phrase,
  nci_code_cui_map_table = "nci_code_cui_map_201808"
)
}
\description{
Query NCI metathesaurus for all exact phrase matches
}
