% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nci_like_phrase.R
\name{query_nci_like_phrase}
\alias{query_nci_like_phrase}
\title{Query NCI for a phrase using a LIKE statement}
\usage{
query_nci_like_phrase(
  nci_phrase,
  nci_code_cui_map_table = "nci_code_cui_map_201808"
)
}
\description{
Query NCI for a phrase using a LIKE statement
}
