% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_phrase_vector.R
\name{query_phrase_vector}
\alias{query_phrase_vector}
\title{Filters MRCONSO table for STR values that exactly match values in the argument vector}
\usage{
query_phrase_vector(vector, limit_per_phrase = NULL)
}
\arguments{
\item{vector}{vector of strings for IN statement in SQL query}
}
\value{
resultSet where STR IN vector argument
}
\description{
Filters MRCONSO table for STR values that exactly match values in the argument vector
}
