% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_str_exact_match_in_mrconso.R
\name{query_str_exact_match_in_mrconso}
\alias{query_str_exact_match_in_mrconso}
\title{Query based on search terms}
\usage{
query_str_exact_match_in_mrconso(phrase, limit = 1)
}
\description{
Query based on search terms
}
