% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_word.R
\name{query_word}
\alias{query_word}
\title{Search for a Word in the MRXW Table}
\usage{
query_word(conn, schema, word)
}
\arguments{
\item{conn}{Database connection}

\item{schema}{Metathesaurus schema}

\item{word}{Word to search}
}
\description{
Unlike the other Search functions that take a `concept` as an argument, this function looks up words and returns all possible matches by CUI and SUI from MRCONSO. More information is available at https://www.nlm.nih.gov/research/umls/implementation_resources/query_diagrams/er6.html.
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}
}
