% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_words.R
\name{query_words}
\alias{query_words}
\title{Query MRXW_ENG table for words and return matching CUI and SUI values in MRCONSO}
\usage{
query_words(..., type = c("or", "and"))
}
\arguments{
\item{...}{words of length 1 each}

\item{type}{"or" if words are to be considered separate queries (as in the case of synonyms); "and" if words are to be combined into a phrase}
}
\description{
Query MRXW_ENG table for words and return matching CUI and SUI values in MRCONSO
}
