% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCUI.R
\name{queryCUI}
\alias{queryCUI}
\title{Queries cui in UMLS in MRCONSO table of metathesaurus}
\usage{
queryCUI(cui, limit_n = NULL)
}
\arguments{
\item{cui}{C-prefixed CUI from metathesaurus}

\item{limit}{LIMIT in the sql statement. If NULL, all records will be returned.}
}
\value{
resultset from SQL query
}
\description{
Queries cui in UMLS in MRCONSO table of metathesaurus
}
