% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_all_possible_cuis.R
\name{return_all_possible_cuis}
\alias{return_all_possible_cuis}
\title{Returns all possible atoms associated with a single phrase search, either like or exact}
\usage{
return_all_possible_cuis(
  phrase,
  search_type = c("like", "exact"),
  limit = NULL
)
}
\arguments{
\item{phrase}{phrase to be searched}

\item{search_type}{like for double-ended wildcards in the SQL statement; exact for an exact match}
}
\description{
Returns all possible atoms associated with a single phrase search, either like or exact
}
