% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_all_semantic_types.R
\name{return_all_semantic_types}
\alias{return_all_semantic_types}
\title{Get all unique semantic types in Metathesaurus}
\usage{
return_all_semantic_types()
}
\description{
Get all unique semantic types in Metathesaurus
}
