% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_valueset_functions.R
\name{return_valueset}
\alias{return_valueset}
\title{Return Valueset}
\usage{
return_valueset(conn, schema = "mth", table, field)
}
\arguments{
\item{conn}{Postgres Connection}

\item{schema}{Schema of MTH instance}

\item{table}{MTH Table}

\item{field}{Field}
}
\description{
Return a valueset to see what the possible values are in a given Field
}
\seealso{
\code{\link[pg13]{query}}
 \code{\link[SqlRender]{render}}
}
