% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_valueset.R
\name{return_valueset_functions}
\alias{return_valueset_functions}
\title{Return the Valueset for a Field}
\arguments{
\item{conn}{Postgres Connection}

\item{schema}{Schema of MTH instance}

\item{table}{MTH Table}

\item{field}{Field}
}
\value{
Vector of 1 or more values
}
\description{
Return a valueset to see what the possible values are in a given Field
}
\keyword{internal}
