% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_scraping_functions.R
\name{scrape_vocabulary_annotations}
\alias{scrape_vocabulary_annotations}
\title{Scrape Vocabulary Annotations}
\usage{
scrape_vocabulary_annotations()
}
\description{
These functions scrape the UMLS Web Documentation to write the \code{\link{mth_documentation_functions}}. They have been kept to enable reproducibility and other long-term maintenance efforts.
}
\keyword{internal}
