% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cui.R
\name{search_cui}
\alias{search_cui}
\title{Gets the most current information about a CUI via API call}
\usage{
search_cui(cui)
}
\arguments{
\item{cui}{CUI to query}
}
\value{
response object or an Error message in console
}
\description{
Gets the most current information about a CUI via API call
}
