% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconso_search_functions.R
\name{search_exact}
\alias{search_exact}
\title{Search for an Exact Match}
\usage{
search_exact(
  conn,
  concept,
  schema = "mth",
  match_case = TRUE,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE
)
}
\description{
Search for an Exact Match
}
\seealso{
\code{\link[SqlRender]{render}}
 \code{\link[pg13]{query}}

Other search mrconso: 
\code{\link{search_like}()},
\code{\link{search_string}()}
}
\concept{search mrconso}
