% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconso_search_functions.R
\name{search_like}
\alias{search_like}
\title{Search for a Like Match}
\usage{
search_like(
  conn,
  concept,
  schema = "mth",
  match_case = TRUE,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE
)
}
\description{
Search for a Like Match
}
\details{
A 'Like' match is one where there is match for a search where there is a wildcard added on both ends of the `concept`
}
\seealso{
\code{\link[SqlRender]{render}}
 \code{\link[pg13]{query}}

Other search mrconso: 
\code{\link{search_exact}()},
\code{\link{search_string}()}
}
\concept{search mrconso}
