% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchCUI.R
\name{searchCUI}
\alias{searchCUI}
\title{Gets the most current information about a CUI via API call}
\usage{
searchCUI(cui)
}
\arguments{
\item{cui}{CUI to query}
}
\value{
response object or an Error message in console
}
\description{
Gets the most current information about a CUI via API call
}
