% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitQuery.R
\name{submitQuery}
\alias{submitQuery}
\title{Submit Query for data retrieval from local umls}
\usage{
submitQuery(sql_statement, ...)
}
\arguments{
\item{...}{Additional arguments passed to DBI's dbGetQuery function}
}
\description{
Submit Query for data retrieval from local umls
}
