% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_tasks.R
\name{tw_get_tasks}
\alias{tw_get_tasks}
\title{Integrate personal and internal boards in Teamwork}
\usage{
tw_get_tasks(projectid)
}
\arguments{
\item{projectid}{"487470" for Meeracle Worker, "487152" for Internal Board}
}
\description{
Integrate personal and internal boards in Teamwork
}
