% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_get_cui.R
\name{umls_api_get_cui}
\alias{umls_api_get_cui}
\title{Gets the most current information about a CUI via API call}
\usage{
umls_api_get_cui(cui)
}
\arguments{
\item{cui}{CUI to query}
}
\value{
response object or an Error message in console
}
\description{
Gets the most current information about a CUI via API call
}
