% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_parse_response.R
\name{umls_api_parse_response}
\alias{umls_api_parse_response}
\title{Takes the response object from the UMLS API Call and parses it into a dataframe}
\usage{
umls_api_parse_response(umls_api_response)
}
\arguments{
\item{umls_api_response}{response object from a call to the UMLS API}
}
\description{
Takes the response object from the UMLS API Call and parses it into a dataframe
}
