% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls_api_search.R
\name{umls_api_search}
\alias{umls_api_search}
\title{UMLS API Search of a String}
\usage{
umls_api_search(string)
}
\arguments{
\item{string}{string to search of length 1}
}
\value{
a list of dataframes from one of each of the following search types: "exact","words","leftTruncation", "rightTruncation", "approximate", "normalizedString", "normalizedWords"
}
\description{
searchType Parameter Behaviors
words: breaks a search term into its component parts, or words, and retrieves all concepts containing any of those words. For example: If you enter “Heart Disease, Acute” a Word search will retrieve all concepts containing any of the three words (heart, or disease, or acute).
approximate: applies lexical variant generation (LVG) rules to the search term and generally results in expanded retrieval of concepts. For example, a search for the term “cold” retrieves all concepts that contain any of the following words: COLDs, chronic obstructive lung disease, chronic obstructive lung diseases, cold, colder, coldest.
exact: retrieves only concepts that include a synonym that exactly matches the search term.
Normalized String: use with English language terms only. Removes lexical variations such as plural and upper case text and compares search terms to the Metathesaurus normalized string index to retrieve relevant concepts. normalizedWords: use with English language terms only. Removes lexical variations such as plural and upper case text, and compares search terms to the Metathesaurus normalized word index to retrieve relevant concepts.
rightTruncation: retrieves concepts with synonyms that begin with the letters of the search term. For example, a right truncation search for “bronch” retrieves concepts that contain synonyms such as bronchitis, bronchiole, bronchial artery.
leftTruncation: retrieves concepts with synonyms that end with the letters of the search term. For example, a left truncation search for “itis” retrieves concepts that contain synonyms such as colitis, bronchitis, pancreatitis.
}
