% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mth_documentation_functions.R
\name{view_readable_table_names}
\alias{view_readable_table_names}
\title{List the Human Readable Names of MTH Tables}
\usage{
view_readable_table_names()
}
\description{
List the Human Readable Names of MTH Tables
}
\seealso{
\code{\link[tibble]{tribble}}
}
