% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{annotate_field_values}
\alias{annotate_field_values}
\title{Annotation Field Values}
\usage{
annotate_field_values(data, ...)
}
\arguments{
\item{...}{(optional) Columns in the data to annotate. If missing, all available annotations are provided for every column.}
}
\description{
Annotation Field Values
}
\details{
These annotations are derived from \code{\link{get_field_value_annotations}}
}
\seealso{
\code{\link[dplyr]{mutate_all}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}}
 \code{\link[purrr]{reduce}}
}
