% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_sab}
\alias{annotate_sab}
\title{Annotate the SAB Field}
\usage{
annotate_sab(data)
}
\arguments{
\item{data}{dataframe returned from a Metathesaurus query}
}
\description{
These functions add extra fields to a dataframe that annotate the field value meaning with what is provided in the UMLS Web Documentation.
}
\details{
The annotations used in this function are derived from \code{\link{get_sab_annotation}}.
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{mutate-joins}}
}
