% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cui_attributes.R,
%   R/query_cui_inverse_relationships.R
\name{query_cui_attributes}
\alias{query_cui_attributes}
\title{Query for source definition}
\usage{
query_cui_attributes(cui)

query_cui_attributes(cui)
}
\description{
Query for source definition

Query for source definition
}
