% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_word.R
\name{query_word}
\alias{query_word}
\title{Query MRXW_ENG table for word and return matching CUI and SUI values in MRCONSO}
\usage{
query_word(word)
}
\arguments{
\item{word}{phrases will return 0.}
}
\description{
Query MRXW_ENG table for word and return matching CUI and SUI values in MRCONSO
}
