% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconso_search_functions.R
\name{search_exact}
\alias{search_exact}
\title{Search for an Exact Match}
\usage{
search_exact(
  conn,
  concept,
  schema = "mth",
  match_case = TRUE,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE
)
}
\arguments{
\item{conn}{Postgres connection}

\item{concept}{Concept to be searched}

\item{schema}{Metathesaurus schema}

\item{match_case}{If TRUE, both the MRCONSO STR column and the `concept` argument are converted to lowercase.}

\item{cache_only}{If TRUE, only the cache is loaded and database querying is skipped. If a cache did not exist for the query, a NULL will be returned.}

\item{skip_cache}{If TRUE, the cache is skipped altogether. The database is directly queried and the resultset is not cached.}

\item{override_cache}{If TRUE, the cache is not loaded, the database is directly queried and the resultset is cached, overwriting any existing cache that may have been present.}
}
\value{
Dataframe of 0 or more rows of all the fields in the MRCONSO Table.
}
\description{
These functions help search for mappings between an a concept and a Metathesaurus identifier in the MRCONSO Table. All queries are cached using the R.cache package for quicker retrieval of the more computationally expensive searches.
}
\seealso{
\code{\link[SqlRender]{render}}
 \code{\link[pg13]{query}}

Other search mrconso: 
\code{\link{mrconso_search_functions}},
\code{\link{search_like}()},
\code{\link{search_string}()}
}
\concept{search mrconso}
