% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconso_search_functions.R
\name{search_string}
\alias{search_string}
\title{Search for a Like Match}
\usage{
search_string(
  conn,
  concept,
  split = " ",
  schema = "mth",
  match_case = TRUE,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE
)
}
\arguments{
\item{split}{Regular expression to split the concept on to generate words for this style of search. Defaults to a space.}
}
\description{
Search for a Like Match
}
\details{
A 'String' style of search splits the concept into a vector of length 1 or greater based on the `split` argument. Seaches of this style will return all matches where all the results of the split were present.
}
\seealso{
\code{\link[pg13]{buildQueryString}},\code{\link[pg13]{query}}

Other search mrconso: 
\code{\link{search_exact}()},
\code{\link{search_like}()}
}
\concept{search mrconso}
