% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconso_search_functions.R
\name{search_string}
\alias{search_string}
\title{Search for a Like Match}
\usage{
search_string(
  conn,
  concept,
  split = " ",
  schema = "mth",
  match_case = TRUE,
  skip_cache = FALSE,
  override_cache = FALSE
)
}
\arguments{
\item{conn}{Postgres connection}

\item{concept}{Concept to be searched}

\item{split}{Regular expression to split the concept on to generate words for this style of search. Defaults to a space.}

\item{schema}{Metathesaurus schema}

\item{match_case}{If TRUE, both the MRCONSO STR column and the `concept` argument are converted to lowercase.}

\item{skip_cache}{If TRUE, the cache is skipped altogether. The database is directly queried and the resultset is not cached.}

\item{override_cache}{If TRUE, the cache is not loaded, the database is directly queried and the resultset is cached, overwriting any existing cache that may have been present.}
}
\value{
Dataframe of 0 or more rows
}
\description{
These functions help search for mappings between an a concept and a Metathesaurus identifier in the MRCONSO Table. All queries are cached using the R.cache package for quicker retrieval of the more computationally expensive searches.
}
\details{
A 'String' style of search splits the concept into a vector of length 1 or greater based on the `split` argument. Seaches of this style will return all matches where all the results of the split were present.
}
\seealso{
\code{\link[pg13]{buildQueryString}},\code{\link[pg13]{query}}

Other search mrconso: 
\code{\link{mrconso_search_functions}},
\code{\link{search_exact}()},
\code{\link{search_like}()}
}
\concept{search mrconso}
